+++ src/abi/structures/abi_route.py
@@ -8,7 +8,7 @@ class ABIRoute(ctypes.Structure):
         ("vehicle_id", ctypes.c_uint32),
         ("stop_ids", ctypes.POINTER(ctypes.c_uint32)),
         ("number_of_stops", ctypes.c_size_t),
-        ("total_distance", ctypes.c_double),
+        ("total_distance", ctypes.c_float),
     ]
 
     def __init__(self, number_of_stops: int, *args, **kw) -> None:
